Combine 3x validation rules in to validation.ts
